build(sinks): feature-gate delta/iceberg/duckdb and slim CI/CD
The DuckDB sink compiles a bundled native (C++) amalgamation, which pushed
`cargo test` from ~15min to ~40min and broke the release pipeline's `cross`/
musl and manylinux/musllinux builds (and would exceed PyPI's 100MB wheel
limit). Make every accepted sink an opt-in Cargo feature so unrelated builds
never pay for a sink they don't use.

floe-core:
- `deltalake`, `iceberg*`, `aws-sdk-glue`, and `duckdb` are now `optional`,
  behind `delta` / `iceberg` / `duckdb` features. Default = ["delta","iceberg"]
  so only the bundled duckdb build is opt-in.
- Gate sink modules, the SINK_FORMATS registry, and the arrow-convert / merge /
  unique-seed helpers with `#[cfg(feature = ...)]`. A config requesting a
  disabled sink fails with a clear "rebuild with --features X" error.
- Relocate generic helpers out of gated modules so non-sink code still builds:
  record-batch builders -> arrow_convert, default_schema_evolution_summary ->
  merge::keys, is_motherduck_connection -> crate::config. MotherDuck handling is
  preserved (additive, not removed).
- Re-export the duckdb crate as `duckdb_driver` (cfg duckdb) so the integration
  test can drop the always-compiled dev-dependency bundle.

floe-cli / floe-python:
- `default-features = false` on the floe-core dep + forwarding features; default
  = ["delta","iceberg"] (duckdb opt-in) to keep `cargo install` and published
  wheels/binaries lean.

CI (ci.yml):
- Lean clippy runs `--no-default-features --features delta,iceberg`; a dedicated
  path-gated `duckdb` job (own cache key) compiles the bundle and runs duckdb
  clippy/tests only when duckdb code changes.

CD (release.yml):
- `build` and `build-python` pass explicit lean features so pre-built binaries
  and wheels never compile the duckdb bundle. DuckDB stays available via source
  build with `--features duckdb`.
