+++ crates/floe-core/Cargo.toml
@@ -17,7 +17,7 @@ yaml-rust2 = "0.11"
 polars = { version = "0.52.0", features = ["csv", "parquet", "lazy", "new_streaming", "timezones", "dtype-date", "dtype-datetime", "dtype-time", "polars-ops", "is_unique", "is_first_distinct"] }
 calamine = "0.24"
 rayon = "1"
-deltalake = { version = "0.30.1", features = ["datafusion", "s3", "azure", "gcs"] }
+deltalake = { version = "0.30.1", features = ["datafusion", "s3", "azure", "gcs"], optional = true }
 url = "2"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
@@ -27,7 +27,8 @@ glob = "0.3"
 roxmltree = "0.20"
 aws-config = "1"
 aws-sdk-s3 = "1"
-aws-sdk-glue = "1"
+# Only used by the Iceberg Glue catalog path; gated under the `iceberg` feature.
+aws-sdk-glue = { version = "1", optional = true }
 apache-avro = "0.16"
 google-cloud-storage = { package = "gcloud-storage", version = "1.2.0" }
 tempfile = "3"
@@ -39,15 +40,15 @@ azure_storage_blobs = "0.20"
 futures = "0.3"
 uuid = "1"
 arrow = "57"
-iceberg = "0.9.0"
-iceberg-catalog-rest = "0.9.0"
-iceberg-storage-opendal = { version = "0.9.1", features = ["opendal-s3", "opendal-gcs"] }
+iceberg = { version = "0.9.0", optional = true }
+iceberg-catalog-rest = { version = "0.9.0", optional = true }
+iceberg-storage-opendal = { version = "0.9.1", features = ["opendal-s3", "opendal-gcs"], optional = true }
 df-interchange = { version = "0.3.2", features = ["arrow_57", "polars_0_52"] }
 orc-rust = "0.7.1"
 # Pinned exactly: 1.10500.0 bundles DuckDB 1.5.0 (native MERGE INTO) and depends on
 # arrow ^57, matching floe-core's `arrow = "57"`. 1.10501+ moved to arrow 58, which would
 # split the RecordBatch type and break `dataframe_to_record_batch` ingestion.
-duckdb = { version = "=1.10500.0", features = ["bundled", "appender-arrow"] }
+duckdb = { version = "=1.10500.0", features = ["bundled", "appender-arrow"], optional = true }
 
 reqwest = { version = "0.12", default-features = true, features = ["native-tls-vendored", "json", "blocking"] }
 sha2 = "0.10"
@@ -57,9 +58,20 @@ thiserror = "1"
 [dev-dependencies]
 rust_xlsxwriter = "0.67"
 mockito = "1"
-# Read-back verification for the DuckDB sink integration tests. Same pin as the
-# main dependency so the test crate reuses the compiled artifact (Arrow 57).
-duckdb = { version = "=1.10500.0", features = ["bundled"] }
 
 [features]
+# `duckdb` is the only sink that pulls a bundled native (C++) build; leaving it
+# always-on roughly tripled full `cargo test` time (~15min → ~40min). Making it
+# opt-in restores the fast default test run. `delta`/`iceberg` are pure-Rust and
+# compile quickly, and the existing validation/manifest/run test suite exercises
+# them through always-compiled config validation (which consults the sink
+# registry), so they stay in the default set to keep `cargo test` green without
+# per-test feature gating. A truly lean consumer can still drop everything with
+# `--no-default-features`. Sink read-back in the DuckDB integration tests reuses
+# the gated crate via the lib re-export (no dev-dep), so `--features duckdb`
+# enables both the writer and its tests.
+default = ["delta", "iceberg"]
+delta = ["dep:deltalake"]
+iceberg = ["dep:iceberg", "dep:iceberg-catalog-rest", "dep:iceberg-storage-opendal", "dep:aws-sdk-glue"]
+duckdb = ["dep:duckdb"]
 vendored-openssl = []
