+++ crates/floe-core/src/io/write/arrow_convert.rs
@@ -5,12 +5,118 @@ use arrow::array::{
     Int64Array, Int8Array, NullArray, StringArray, Time64MicrosecondArray, Time64NanosecondArray,
     TimestampMicrosecondArray, UInt16Array, UInt32Array, UInt64Array, UInt8Array,
 };
+#[cfg(any(feature = "delta", feature = "duckdb"))]
+use arrow::datatypes::{Field, Schema};
+#[cfg(any(feature = "delta", feature = "duckdb"))]
+use arrow::record_batch::RecordBatch;
+#[cfg(any(feature = "delta", feature = "duckdb"))]
+use polars::prelude::DataFrame;
 use polars::prelude::{DataType, Series, TimeUnit};
 
+#[cfg(any(feature = "delta", feature = "duckdb"))]
+use crate::checks::normalize;
+#[cfg(any(feature = "delta", feature = "duckdb"))]
+use crate::config;
 use crate::errors::RunError;
 use crate::FloeResult;
 
+/// Convert a Polars `DataFrame` into an Arrow `RecordBatch`, honoring the entity's
+/// declared output schema when present (column selection, renaming, nullability).
+/// Shared by the Delta and DuckDB sinks; gated to those features.
+#[cfg(any(feature = "delta", feature = "duckdb"))]
+pub(crate) fn dataframe_to_record_batch(
+    df: &DataFrame,
+    entity: &config::EntityConfig,
+) -> FloeResult<RecordBatch> {
+    if entity.schema.columns.is_empty() {
+        return dataframe_to_record_batch_all(df);
+    }
+
+    let schema_columns = normalize::resolve_output_columns(
+        &entity.schema.columns,
+        normalize::resolve_normalize_strategy(entity)?.as_deref(),
+    );
+    dataframe_to_record_batch_with_schema(df, &schema_columns)
+}
+
+#[cfg(any(feature = "delta", feature = "duckdb"))]
+pub(crate) fn dataframe_to_record_batch_with_schema(
+    df: &DataFrame,
+    schema_columns: &[config::ColumnConfig],
+) -> FloeResult<RecordBatch> {
+    let mut fields = Vec::with_capacity(schema_columns.len());
+    let mut arrays = Vec::with_capacity(schema_columns.len());
+    for column in schema_columns {
+        let series = df
+            .column(column.name.as_str())
+            .map_err(|err| Box::new(RunError(format!("column lookup failed: {err}"))))?;
+        let series = series.as_materialized_series();
+        let array = record_batch_series_to_arrow(series)?;
+        let nullable = column.nullable.unwrap_or(true);
+        if !nullable && array.null_count() > 0 {
+            return Err(Box::new(RunError(format!(
+                "write rejected nulls for non-nullable column {}",
+                column.name
+            ))));
+        }
+        fields.push(Field::new(
+            column.name.clone(),
+            array.data_type().clone(),
+            nullable,
+        ));
+        arrays.push(array);
+    }
+    let schema = Arc::new(Schema::new(fields));
+    RecordBatch::try_new(schema, arrays).map_err(|err| {
+        Box::new(RunError(format!("record batch build failed: {err}")))
+            as Box<dyn std::error::Error + Send + Sync>
+    })
+}
+
+#[cfg(any(feature = "delta", feature = "duckdb"))]
+pub(crate) fn dataframe_to_record_batch_all(df: &DataFrame) -> FloeResult<RecordBatch> {
+    let mut fields = Vec::with_capacity(df.width());
+    let mut arrays = Vec::with_capacity(df.width());
+    for column in df.get_columns() {
+        let series = column.as_materialized_series();
+        let name = series.name().to_string();
+        let array = record_batch_series_to_arrow(series)?;
+        let nullable = array.null_count() > 0;
+        fields.push(Field::new(name, array.data_type().clone(), nullable));
+        arrays.push(array);
+    }
+    let schema = Arc::new(Schema::new(fields));
+    RecordBatch::try_new(schema, arrays).map_err(|err| {
+        Box::new(RunError(format!("record batch build failed: {err}")))
+            as Box<dyn std::error::Error + Send + Sync>
+    })
+}
+
+#[cfg(any(feature = "delta", feature = "duckdb"))]
+fn record_batch_series_to_arrow(series: &Series) -> FloeResult<ArrayRef> {
+    series_to_arrow_array(
+        series,
+        ArrowConversionOptions {
+            upcast_i8_i16_to_i32: false,
+            time_encoding: ArrowTimeEncoding::Nanoseconds,
+        },
+        |dtype| {
+            RunError(format!(
+                "sink does not support dtype {dtype:?} for {}",
+                series.name()
+            ))
+        },
+    )
+}
+
+// Each variant is used by a different sink (Nanoseconds → Delta/DuckDB record
+// batches, Microseconds → Iceberg), so single-feature builds legitimately leave
+// one variant unconstructed.
 #[derive(Debug, Clone, Copy)]
+#[cfg_attr(
+    not(all(any(feature = "delta", feature = "duckdb"), feature = "iceberg")),
+    allow(dead_code)
+)]
 pub(crate) enum ArrowTimeEncoding {
     Nanoseconds,
     Microseconds,
