+++ crates/floe-core/tests/integration/mod.rs
@@ -1,12 +1,18 @@
 pub mod archive_run;
 pub mod composite_unique;
+#[cfg(feature = "delta")]
 pub mod delta_run;
 pub mod dry_run;
+#[cfg(feature = "duckdb")]
 pub mod duckdb_run;
 pub mod fixed_width;
+#[cfg(feature = "iceberg")]
 pub mod iceberg_gcs_run;
+#[cfg(feature = "iceberg")]
 pub mod iceberg_glue_run;
+#[cfg(feature = "iceberg")]
 pub mod iceberg_run;
+#[cfg(feature = "iceberg")]
 pub mod iceberg_s3_run;
 pub mod json_selectors;
 pub mod local_run;
