+++ crates/floe-core/tests/unit/io/write/mod.rs
@@ -1,7 +1,11 @@
+#[cfg(feature = "delta")]
 pub mod delta_merge;
+#[cfg(feature = "delta")]
 pub mod delta_write;
+#[cfg(feature = "iceberg")]
 pub mod iceberg_write;
 pub mod metrics;
+#[cfg(all(feature = "delta", feature = "iceberg"))]
 pub mod object_store;
 pub mod parquet_write;
 pub mod parts;
